requirements.txt
asn1crypto==0.24.0
attributee==0.1.3
attrs==21.2.0
bidict==0.21.2
cachetools==4.2.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.4.4
conda==4.5.4
cryptography==2.2.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dominate==2.6.0
future==0.18.2
idna==2.6
imageio==2.9.0
importlib-metadata==4.0.1
joblib==1.0.1
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
networkx==2.5.1
nibabel==3.2.1
numba==0.53.1
numpy==1.19.5
opencv-python==4.1.2.30
ordered-set==4.0.2
packaging==20.9
pandas==1.1.5
phx-class-registry==3.0.5
Pillow==8.1.2
pycosat==0.6.3
pycparser==2.18
PyLaTeX==1.4.1
pyOpenSSL==18.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.6.8
python-dateutil==2.8.1
pytorch-metric-learning==0.9.99
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
ruamel-yaml==0.15.37
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.11.0
spatial-correlation-sampler==0.3.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.9.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.22
visdom==0.1.8.9
vot-trax==3.0.3
websocket-client==1.0.1
zipp==3.4.1
